backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.0
pydantic==2.7.1
pydantic-settings==2.3.0
sqlalchemy==2.0.30
aiosqlite==0.20.0
alembic==1.13.1
firebase-admin==6.5.0
python-multipart==0.0.9
slowapi==0.1.9
